pyproject.toml
# Copyright (c) 2024, NVIDIA CORPORATION.  All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[tool.isort]
profile = "black"  # black-compatible
line_length = 119  # should match black parameters
py_version = 310  # python 3.10 as a target version


[tool.black]
line_length = 119
skip_string_normalization = true

[project]
name = "nemo_skills"
description = "NeMo Skills - a project to improve skills of LLMs"
dynamic = ["version"]
requires-python = ">= 3.10"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
  'hydra-core',
  'tqdm',
  'pyyaml',
  'numpy',
  'requests',
  'backoff',
  'sympy',
  'rank_bm25',
  'openai',
]

[tool.setuptools.packages.find]
include = ["nemo_skills*"]

[tool.setuptools.dynamic]
version = {attr = "nemo_skills.__version__"}
